Add files via upload
 Problem Set #1 - Linear Regression
This repository contains Python implementations and solutions for the exercises outlined in Problem Set #1: Linear Regression . The goal of this assignment is to enhance your understanding of linear regression and its extensions through hands-on coding and visualization. Each question addresses a specific aspect of linear regression, such as model training, evaluation, basis functions, weighted regression, and more.